Factor out LsdfOpType.js
